Update and cleanup
- added reaction for timer reaching 0 sec.
- fixed issue with memory leak
- fixed issue with onBackPressed bug preventing from exiting the app
- implemented factory pattern in addingNewTimeModeFragment
- added/improved compatibility for different screen sizes

+++ app/src/main/java/com/example/chesstimer/addingNewTimeModeFragment/timeChanger/HourChanger.kt
@@ -1,9 +1,9 @@
 package com.example.chesstimer.addingNewTimeModeFragment.timeChanger
 
-class HourChanger:TimeChanger() {
-    override val minInt: Long
+class HourChanger : TimeChanger() {
+    override val minInt: Int
         get() = 0
-    override val maxInt: Long
+    override val maxInt: Int
         get() = 99
     override val timeInterval: Int
         get() = 1
