+++ app/src/main/java/com/example/chesstimer/addingNewTimeModeFragment/timeChanger/MinuteChanger.kt
@@ -1,9 +1,9 @@
 package com.example.chesstimer.addingNewTimeModeFragment.timeChanger
 
-class MinuteChanger: TimeChanger() {
-    override val minInt: Long
+class MinuteChanger : TimeChanger() {
+    override val minInt: Int
         get() = 0
-    override val maxInt: Long
+    override val maxInt: Int
         get() = 59
     override val timeInterval: Int
         get() = 1
