+++ app/src/main/java/com/example/chesstimer/addingNewTimeModeFragment/timeChanger/SecondChanger.kt
@@ -1,9 +1,9 @@
 package com.example.chesstimer.addingNewTimeModeFragment.timeChanger
 
-class SecondChanger:TimeChanger() {
-    override val minInt: Long
+class SecondChanger : TimeChanger() {
+    override val minInt: Int
         get() = 0
-    override val maxInt: Long
+    override val maxInt: Int
         get() = 50
     override val timeInterval: Int
         get() = 10
